Add files via upload
This represents work done in Windows (actually WSL Fedora Linux "Remix", the only place in Windows where I can use gforth), on 12/8/23. The work should represent some legitimate advances from the "main" branch, but there are still some kinks to work out. "Read-nt" crashes when given a non-valid note number to display. This is fallout from labeling all the notes at init-time as "unused" by adding the dummy code byte hex F5 as the index to all the new, unwritten notes.

Deleted notes have the "F4" code as their index, and "F4" is used as a sort of "null pointer" to indicate the end of the linked list of notes. "F4" is being used in the queue of notes available to signify note numbers that have been "dequeued" - given for use to the main notes.fs program for each new note (see "write-nt").

If I can debug this, it should probably be remerged with the "main" branch.