An entirely new approach, using a linked list
Among other developments, I decided to use a linked list of structs to hold notes. The notes can now be more than one line; up to 256 characters. 

The notes will be serialized to be written to a file. Obviously the linked list will be reconstructed from the serialized data file.

I have added date/timestamps to indicate when the note was written.

Other changes will be reported as they are made.